+++ src/main/java/core/gui/comp/renderer/HODateTimeTableCellRenderer.java
@@ -1,24 +1,17 @@
 package core.gui.comp.renderer;
-
+import core.util.HODateTime;
 import java.awt.Component;
-import java.util.Date;
-import java.text.DateFormat;
-
 import javax.swing.JTable;
 import javax.swing.table.DefaultTableCellRenderer;
 
-public class DateTimeTableCellRenderer extends DefaultTableCellRenderer {
-
-	private static final long serialVersionUID = -5869341433817862361L;
-	private DateFormat format = DateFormat.getDateTimeInstance(DateFormat.MEDIUM, DateFormat.SHORT);
-
+public class HODateTimeTableCellRenderer extends DefaultTableCellRenderer {
 	@Override
 	public Component getTableCellRendererComponent(JTable table, Object value, boolean isSelected,
 			boolean hasFocus, int row, int column) {
 
 		String dateString = "";
 		if (value != null) {
-			dateString = this.format.format((Date) value);
+			dateString = ((HODateTime) value).toLocaleDateTime();
 		}
 		return super.getTableCellRendererComponent(table, dateString, isSelected, hasFocus, row,
 				column);