+++ src/main/java/core/model/player/Player.java
@@ -333,7 +333,7 @@ public class Player extends AbstractTable.Storable {
     private Integer motherclubId;
     private String motherclubName;
     private Integer matchesCurrentTeam;
-    private int hrf_id;
+    private int hrf_id = -1;
     private Integer htms = null;
     private Integer htms28 = null;
 
@@ -612,13 +612,14 @@ public class Player extends AbstractTable.Storable {
     /**
      * Calculates String for full age and days correcting for the difference between (now and last HRF file)
      *
-     * @return String of age & agedays format is "YY (DDD)"
+     * @return String of age & age days format is "YY (DDD)"
      */
     public String getAgeWithDaysAsString() {
         return getAgeWithDaysAsString(HODateTime.now());
     }
     public String getAgeWithDaysAsString(HODateTime t){
-        return getAgeWithDaysAsString(this.getAlter(), this.getAgeDays(), t, this.m_clhrfDate);
+        if (this.m_clhrfDate != null) return getAgeWithDaysAsString(this.getAlter(), this.getAgeDays(), t, this.m_clhrfDate);
+        return "";
     }
 
     /**
@@ -643,6 +644,11 @@ public class Player extends AbstractTable.Storable {
     public static String getAgeWithDaysAsString(int ageYears, int ageDays, HODateTime time, HODateTime hrfTime) {
         var age = new HODateTime.HODuration(ageYears, ageDays).plus(HODateTime.HODuration.between(hrfTime, time));
         return age.seasons + " (" + age.days + ")";
+    }
+
+    public HODateTime.HODuration getAgeAtDate(HODateTime date){
+        if ( this.m_clhrfDate != null) return new HODateTime.HODuration(this.getAlter(), this.getAgeDays()).plus(HODateTime.HODuration.between(this.m_clhrfDate, date));
+        return null;
     }
 
     /**
@@ -1256,15 +1262,16 @@ public class Player extends AbstractTable.Storable {
 
     }
 
-
     public String getFullName() {
 
-        if (getNickName().isEmpty())
-        {
-            return getFirstName() + " " +getLastName();
+        if (getNickName().isEmpty()) {
+            if (!getFirstName().isEmpty()) {
+                return getFirstName() + " " + getLastName();
+            }
+            return getLastName();
         }
 
-        return getFirstName() + " '" + getNickName() + "' " +getLastName();
+        return getFirstName() + " '" + getNickName() + "' " + getLastName();
     }
 
     /**
@@ -1814,6 +1821,7 @@ public class Player extends AbstractTable.Storable {
     public int getHrfId() {
         return this.hrf_id;
     }
+    public boolean isTemporary(){return this.hrf_id == -1;}
 
     public void setLastMatchDate(String v) {
         this.m_lastMatchDate = v;