+++ src/main/java/module/transfer/PlayerTransfer.java
@@ -1,4 +1,3 @@
-// %2277986132:hoplugins.transfers.vo%
 package module.transfer;
 
 import core.db.AbstractTable;
@@ -6,6 +5,13 @@ import core.db.DBManager;
 import core.model.HOVerwaltung;
 import core.model.player.Player;
 import core.util.HODateTime;
+import core.util.HOLogger;
+import java.time.temporal.ChronoUnit;
+import java.util.Comparator;
+import java.util.List;
+
+import static java.lang.Math.min;
+import static module.transfer.XMLParser.downloadTeamTransfers;
 
 /**
  * Value Object representing a player transfer.
@@ -23,6 +29,13 @@ public class PlayerTransfer extends AbstractTable.Storable {
 
     /** Type to indicate a SELL transfer */
     public static final int REBOUGHT = 0;
+
+    public static Integer teamId;
+    public static Long totalSumOfBuys;
+    public static Long totalSumOfSales;
+    public static Long numberOfBuys;
+    public static Long numberOfSales;
+    public static HODateTime activatedDate;
 
     //~ Instance fields ----------------------------------------------------------------------------
 
@@ -44,10 +57,7 @@ public class PlayerTransfer extends AbstractTable.Storable {
     /** Id of the buyer team */
     private int buyerid = 0;
 
-    /** Market value of the player at transfer date */
-    private int marketvalue = 0;
-
-    /** Id of the transfered player */
+    /** Id of the transferred player */
     private int playerId;
 
     /** Transfer price */
@@ -64,6 +74,13 @@ public class PlayerTransfer extends AbstractTable.Storable {
 
     /** TSI value of the player at transfer date */
     private int tsi = 0;
+    
+    private Integer motherClubFee;
+    private Integer previousClubFee;
+
+    public void setType(int type) {
+        this.type = type;
+    }
 
     /** Transfer type */
     private int type;
@@ -88,6 +105,166 @@ public class PlayerTransfer extends AbstractTable.Storable {
      * constructor is used by AbstractTable (TransfersTable)
      */
     public PlayerTransfer(){}
+
+    /**
+     * Download team's missing player transfers
+     * If downloaded transfer is not stored in db or differs from stored version, the database is updated.
+     * @param teamId Team id
+     */
+    public static void downloadMissingTransfers(int teamId) {
+        var db = DBManager.instance();
+        var isSold = true;
+        var sold = db.getSumTransferPrices(teamId, isSold);
+        var bought = db.getSumTransferPrices(teamId, !isSold);
+        int page = 0;
+        do {
+            var transfers = downloadTeamTransfers(teamId, page++);
+            if (transfers.isEmpty()) break;
+            for (var transfer : transfers) {
+                var inDB = db.loadPlayerTransfer(transfer.transferId);
+                if (inDB == null || transfer.getPrice() != inDB.getPrice() || inDB.getPlayerId()==0) {
+                    updatePlayerTransfer(transfer);
+                    var priceCorrection = transfer.getPrice();
+                    if (inDB != null) priceCorrection -= inDB.getPrice();
+                    if (transfer.buyerid == teamId) bought += priceCorrection;
+                    else sold += priceCorrection;
+                }
+            }
+        } while (sold != PlayerTransfer.totalSumOfSales || bought != PlayerTransfer.totalSumOfBuys);
+    }
+
+    /**
+     * Update player transfer in database.
+     * If transfer's player is a deleted/fired player (player id is 0), the correct player id is tried to load from
+     * previously stored version of the transfer. If this is not available, the player info is loaded from player table (SPIELER).
+     * @param transfer Player transfer
+     */
+    private static void updatePlayerTransfer(PlayerTransfer transfer) {
+        DBManager.instance().storePlayerTransfer(transfer);
+    }
+
+    /**
+     * Get the mother club fee
+     * If not available, the value is calculated.
+     * @return int
+     */
+    public int getMotherClubFee(){
+        if ( this.motherClubFee == null) {
+            var teamId = HOVerwaltung.instance().getModel().getBasics().getTeamId();
+            this.motherClubFee = calcMotherClubFee(teamId);
+        }
+        return this.motherClubFee;
+    }
+
+    /**
+     * Set the mother club fee
+     * @param motherClubFee Integer
+     */
+    public void setMotherClubFee(Integer motherClubFee) {
+        this.motherClubFee = motherClubFee;
+    }
+
+    /**
+     * Calculate the mother club fee
+     * @param teamId Team id
+     * @return int
+     */
+    private int calcMotherClubFee(int teamId){
+        var player = getPlayerInfo();
+        if (player != null && player.isHomeGrown() && this.sellerid != teamId) return (int)(this.price * .02);    // 2%
+        return 0;
+    }
+
+    /**
+     * Get the income of transfer's previous club fees
+     * It is calculated, if not available yet.
+     * @return int The income in locale currency
+     */
+    public int getPreviousClubFee(){
+        if ( this.previousClubFee == null) {
+            var teamId = HOVerwaltung.instance().getModel().getBasics().getTeamId();
+            this.previousClubFee = calcPreviousClubFee(teamId);
+        }
+        return this.previousClubFee;
+    }
+
+    public void setPreviousClubFee(Integer motherClubFee) {
+        this.previousClubFee = motherClubFee;
+    }
+
+    /**
+     * Calculate the income of transfer's previous club fees
+     * see <a href="https://www85.hattrick.org/Help/Rules/AppTransferFees.aspx">...</a>
+     * @param teamId Team ID
+     * @return Income in locale currency
+     */
+    private int calcPreviousClubFee(int teamId) {
+        var player = getPlayerInfo();
+        if (player != null) {
+            var matchCount = player.getMatchesCurrentTeam();
+            if (matchCount==null){
+                matchCount = DBManager.instance().getCountOfPlayedMatches(playerId, true);
+                player.setMatchesCurrentTeam(matchCount);
+            }
+            if (matchCount > 0) {
+                // get previous transfer
+                PlayerTransfer previous = null;
+                var transfers = DBManager.instance().getTransfers(getPlayerId(), true);
+                for (var transfer : transfers) {
+                    if (transfer.getDate().isBefore(this.getDate())) {
+                        if (previous == null || previous.getDate().isBefore(transfer.getDate())) {
+                            previous = transfer;
+                        }
+                    }
+                }
+                if (previous != null && previous.getSellerid() == teamId) {
+                    var percentage = switch (matchCount) {
+                        case 1 -> 0.0025;
+                        case 2 -> 0.005;
+                        case 3 -> 0.01;
+                        case 4 -> 0.015;
+                        case 5, 6 -> 0.02;
+                        case 7, 8, 9 -> 0.025;
+                        default -> min(0.04, 0.025 + matchCount / 10 * 0.005);
+                    };
+                    return (int) (percentage * getPrice());
+                }
+            }
+        }
+        return 0;
+    }
+
+    /**
+     * Download transfers that were responsible for commission income of given ht week
+     * Methods examines sum of already stored player transfers. If income is missing,
+     * the list of sold players is loaded and beginning with the latest sold player, their
+     * transfers are downloaded from hattrick. If a missing transfer is found,
+     * it is stored and it's commission is added to the sum of db entries.
+     * The loop is exited, when reaching the commission value.
+     * @param soldPlayers List of sold players by the team
+     * @param commission  Commission income of the week from economy download
+     * @param htweek HT week of that commission income
+     */
+    public static void downloadMissingTransferCommissions(List<PlayerTransfer> soldPlayers, int commission, HODateTime.HTWeek htweek) {
+        var startWeek = HODateTime.fromHTWeek(htweek);
+        var sum = DBManager.instance().getSumTransferCommissions(startWeek);
+        if(sum != commission){
+            for (var player : soldPlayers){
+                var transfers = XMLParser.getAllPlayerTransfers(player.getPlayerId());
+                for (var transfer : transfers){
+                    var inDB = DBManager.instance().loadPlayerTransfer(transfer.getTransferId());
+                    if (inDB == null){
+                        DBManager.instance().storePlayerTransfer(transfer);
+                        if (!transfer.getDate().isBefore(startWeek) && !transfer.getDate().isAfter(startWeek.plus(7, ChronoUnit.DAYS))) {
+                            sum += transfer.getMotherClubFee();
+                            sum += transfer.getPreviousClubFee();
+                            if (sum >= commission) return;
+                        }
+                    }
+                }
+            }
+        }
+    }
 
     //~ Methods ------------------------------------------------------------------------------------
 
@@ -154,24 +331,6 @@ public class PlayerTransfer extends AbstractTable.Storable {
     }
 
     /**
-     * Sets the market value of the player on transfer date.
-     *
-     * @param value Market value of the player on transfer date.
-     */
-    public final void setMarketvalue(Integer value) {
-        if ( value != null ) this.marketvalue = value;
-    }
-
-    /**
-     * Gets the market value of the player on transfer date.
-     *
-     * @return Market value of the player on transfer date.
-     */
-    public final int getMarketvalue() {
-        return marketvalue;
-    }
-
-    /**
      * Sets the id of the transfered player.
      *
      * @param id of the transfered player.
@@ -189,40 +348,143 @@ public class PlayerTransfer extends AbstractTable.Storable {
         return playerId;
     }
 
-//    /**
-//     * Sets the information about the player on transfer date.
-//     *
-//     * @param info Player information on transfer date.
-//     */
-//    public final void setPlayerInfo(Player info) {
-//        this.playerInfo = info;
-//    }
-
     /**
      * Gets the information about the player on transfer date.
      *
-     * @return Playerinformation if available. else <code>null</code>
+     * @return Player information if available. else <code>null</code>
      */
     public final Player getPlayerInfo() {
-        if ( playerInfo == null){
-            playerInfo = DBManager.instance().getSpielerAtDate(this.getPlayerId(), this.getDate().toDbTimestamp());
+        if (playerInfo == null) {
+            playerInfo = loadPLayerInfo(false);
         }
         return playerInfo;
     }
 
     /**
-     * Sets the name of the transfered player.
-     *
-     * @param name Name of the transfered player.
+     * Try to load information about the player at transfer date
+     * If no information is found, a temporary player object is created with limited information got from the transfer record.
+     * @param doUpdate True: Update database if values differ from database contents
+     * @return Player
+     */
+    public Player loadPLayerInfo(boolean doUpdate) {
+        var oldPLayerId = playerId;
+        if (playerId == 0 && !this.isStored()) {
+            // Try to find correct player id in previously stored transfer
+            var transferInDb = DBManager.instance().loadPlayerTransfer(this.transferId);
+            if (transferInDb != null) {
+                playerId = transferInDb.getPlayerId();
+                oldPLayerId = playerId;
+            }
+        }
+        var isPurchase = this.buyerid == HOVerwaltung.instance().getModel().getBasics().getTeamId();
+        if (playerId > 0) {
+            loadPLayerInfoAtTransferDate(isPurchase);
+        } else if (playerId != -1) {
+            HODateTime start = HODateTime.now();
+            List<Player> playerInfos = loadPLayerInfosAtTransferDate(isPurchase);
+            HOLogger.instance().debug(getClass(), this.getPlayerName() + " loaded candidate count: " + playerInfos.size() + " Started: " + start.toLocaleDateTime() + " Duration: " + HODateTime.between(start, HODateTime.now()).toString());
+            if (playerInfos.size() > 1) {
+                // find most probable candidate
+                if (isPurchase) {
+                    playerInfo = playerInfos.stream().min(Comparator.comparing(Player::getHrfDate)).get();
+                } else {
+                    playerInfo = playerInfos.stream().max(Comparator.comparing(Player::getHrfDate)).get();
+                }
+            } else if (playerInfos.size() == 1) {
+                playerInfo = playerInfos.get(0);
+            }
+
+            if (playerInfo != null) {
+                this.playerId = playerInfo.getPlayerID();
+                // Don't update database on inserts (causes recursive inserts on transfer updates)
+                if ( doUpdate ) {
+                    updatePlayerTransfer(this);
+                }
+            }
+            else {
+                playerId = -1; // fired player
+            }
+        }
+        if (playerInfo == null) {
+            playerInfo = new Player();
+            playerInfo.setLastName(this.getPlayerName());
+            playerInfo.setTSI(this.getTsi());
+            playerInfo.setPlayerID(playerId);
+            playerInfo.setLeadership(-1);
+            playerInfo.setExperience(-1);
+            playerInfo.setForm(-1);
+            playerInfo.setStamina(-1);
+            playerInfo.setTorwart(-1);
+            playerInfo.setVerteidigung(-1);
+            playerInfo.setSpielaufbau(-1);
+            playerInfo.setPasspiel(-1);
+            playerInfo.setFluegelspiel(-1);
+            playerInfo.setTorschuss(-1);
+            playerInfo.setStandards(-1);
+
+            if (doUpdate && oldPLayerId != playerId) {
+                // Don't update database on inserts (causes recursive inserts on transfer updates)
+                updatePlayerTransfer(this);
+            }
+        }
+        return playerInfo;
+    }
+
+    /**
+     * Load player candidate records near transfer date by player name.
+     * If it is a purchase the player record downloaded after given date are loaded, otherwise the latest before.
+     * If no records are found, the other time direction is tried too.
+     * @param isPurchase Boolean
+     * @return Since player names are not unique a list of candidates is returned
+     */
+    private List<Player> loadPLayerInfosAtTransferDate(boolean isPurchase) {
+        List<Player>playerInfos;
+        if (isPurchase) {
+            playerInfos = DBManager.instance().getFirstPlayerDownloadAfter(this.getPlayerName(), this.getDate().toDbTimestamp());
+            if ( playerInfos.isEmpty()){
+                playerInfos = DBManager.instance().getLatestPlayerDownloadBefore(this.getPlayerName(), this.getDate().toDbTimestamp());
+            }
+        } else {
+            playerInfos = DBManager.instance().getLatestPlayerDownloadBefore(this.getPlayerName(), this.getDate().toDbTimestamp());
+            if ( playerInfos.isEmpty()){
+                playerInfos = DBManager.instance().getFirstPlayerDownloadAfter(this.getPlayerName(), this.getDate().toDbTimestamp());
+            }
+        }
+        return playerInfos;
+    }
+
+    /**
+     * Load player candidate records near transfer date by player id.
+     * If it is a purchase the first player record downloaded after given date is loaded, otherwise the latest before.
+     * If no records are found, the other time direction is tried too.
+     * @param isPurchase Boolean
+     */
+    private void loadPLayerInfoAtTransferDate(boolean isPurchase) {
+        if (isPurchase) {
+            playerInfo = DBManager.instance().getFirstPlayerDownloadAfter(this.getPlayerId(), this.getDate().toDbTimestamp());
+            if (playerInfo == null){
+                playerInfo= DBManager.instance().getLatestPlayerDownloadBefore(this.getPlayerId(), this.getDate().toDbTimestamp());
+            }
+        } else {
+            // transfers of sold players (sellerid must not be equal to teamid)
+            playerInfo = DBManager.instance().getLatestPlayerDownloadBefore(this.getPlayerId(), this.getDate().toDbTimestamp());
+            if (playerInfo == null){
+                playerInfo= DBManager.instance().getFirstPlayerDownloadAfter(this.getPlayerId(), this.getDate().toDbTimestamp());
+            }
+        }
+    }
+
+    /**
+     * Sets the name of the transferred player.
+     * @param name Name of the transferred player.
      */
     public final void setPlayerName(String name) {
         this.playerName = name;
     }
 
     /**
-     * Gets the name of the transfered player.
-     *
-     * @return Name of the transfered player.
+     * Gets the name of the transferred player.
+     * @return Name of the transferred player.
      */
     public final String getPlayerName() {
         return playerName;
@@ -230,7 +492,7 @@ public class PlayerTransfer extends AbstractTable.Storable {
 
     /**
      * Sets the transfer price.
-     *
+     * Hattrick currency (swedish crone)
      * @param price Transfer price.
      */
     public final void setPrice(Integer price) {
@@ -240,7 +502,7 @@ public class PlayerTransfer extends AbstractTable.Storable {
     /**
      * Gets the tranfer price.
      *
-     * @return Transfer price
+     * @return Transfer price in Hattrick currency (swedish crone)
      */
     public final int getPrice() {
         return price;